% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDiv.R
\name{betaDiv}
\alias{betaDiv}
\title{calculation of beta diversity between all pairs of communities.}
\usage{
betaDiv(x, method = "bray", binary = FALSE, diag = FALSE, upper = FALSE, ...)
}
\arguments{
\item{x}{The community matrix}

\item{method}{The method to calculate beta diversity}

\item{binary}{Whether convert to presence/absence (using decostand) before
calculation}

\item{diag}{Compute diagonals}

\item{upper}{Whether return only the upper diagonal in the matrix}

\item{...}{Other parameters to parse to vegan::vegdist()}
}
\value{
a dissimilarity matrix in a dist object
}
\description{
This is the first step for all calculations. Beta diversity will be
calculated using the vegan::vegdist function.
By default, the method is "bray" (Bray-Curtis dissimilarity)
}
\examples{
library(vegan)
 data(varespec)
View(head(varespec))
example.comdist <- betaDiv(varespec)

}
