% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamPred.R
\name{gamPred}
\alias{gamPred}
\title{calculation of stability using an generalized additive model.}
\usage{
gamPred(
  comtable,
  envmeta,
  comdist = NULL,
  sitenames = NULL,
  GAM.dist.method = "manhattan",
  symmetric = FALSE
)
}
\arguments{
\item{comtable}{The community table}

\item{envmeta}{The environmental metadata table/matrix}

\item{comdist}{The community dissimilarity matrix (optional, default: NULL)}

\item{sitenames}{The names of the site (optional, default: NULL)}

\item{GAM.dist.method}{The method for calculating dist (default: "manhattan")}

\item{symmetric}{Whether to use symmetric algorithm in the calculating.}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the community count table and the environmental
metadata table as input, and calculate the stability of each site using a
generalized additive model (GAM). Alternatively, if no dissimilarity matrix
of the community is provided, the function will calculate the community
dissimilarity based on Bray-Curtis distance and use it. In GAM, the
prediction results are expected to perform better than the linear models.
}
\examples{
library(vegan)
data(varespec)
data(varechem)
example.stability_GAM <- gamPred(varespec, varechem)

}
