% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPred.R
\name{rfPred}
\alias{rfPred}
\title{calculation of stability using a random forest model.}
\usage{
rfPred(comdist, envmeta, sitenames = NULL, seed = NULL, symmetric = FALSE)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envmeta}{The environmental metadata table/matrix}

\item{sitenames}{The names of the site}

\item{seed}{The random seed for reproducibility of the random forest model}

\item{symmetric}{Whether to use symmetric algorithm in the calculating.}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the dissimilarity matrix and the environmental
matrix as input, and calculate the stability of each site using a random
forest model to improve the prediction performance.
}
\examples{
library(vegan)
data(varespec)
data(varechem)
example.comdist <- vegdist(varespec[1:10,])
example.stability_RF <- rfPred(example.comdist, varechem[1:10,])

}
