% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varecoords}
\alias{varecoords}
\title{Coordinates data}
\format{
\subsection{\code{varecoords}}{

A data frame with 28 rows and 3 columns:
\describe{
\item{site}{site name/number}
\item{x}{longitude}
\item{y}{latitude}
...
}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.2307/3236351}
}
\usage{
varecoords
}
\description{
A dataset of coordinate information for vegan::varechem retrieved by
georeferencing the Fig.1 of the original article "Effects of reindeer
grazing on understorey vegetation in dry \if{html}{\out{<i>}}Pinus sylvestris\if{html}{\out{</i>}} forests" by
Väre et al. (1995). NOTICE: The data is not precise nor original.
}
\keyword{datasets}
