% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostPred.R
\name{xgboostPred}
\alias{xgboostPred}
\title{calculation of stability using a xgboost model.}
\usage{
xgboostPred(
  comdist,
  envmeta,
  sitenames = NULL,
  seed = NULL,
  params = NULL,
  symmetric = FALSE
)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envmeta}{The environmental metadata table/matrix}

\item{sitenames}{The names of the site}

\item{seed}{The random seed for reproducibility of the xgboost model}

\item{params}{A list of parameters for the xgboost model.
If NULL, default parameters will be used.}

\item{symmetric}{Whether to use symmetric algorithm in the calculating.}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the dissimilarity matrix and the environmental
matrix as input, and calculate the stability of each site using a xgboost
model to improve the prediction performance.
}
\examples{
library(vegan)
data(varespec)
data(varechem)
example.comdist <- vegdist(varespec[1:10,])
example.stability_XGB <- xgboostPred(example.comdist, varechem[1:10,])

}
