% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStability.R
\name{calcStability}
\alias{calcStability}
\title{calculate stability based on predicted and measured distances.}
\usage{
calcStability(predicted.dist, measured.dist)
}
\arguments{
\item{predicted.dist}{The predicted distance (shall be in range 0~1)}

\item{measured.dist}{The measured distance (shall be in range 0~1)}
}
\value{
a numeric value of stability for the site in range [-1, 1].
}
\description{
This function calculates the stability of a site by comparing the predicted
distance and the measured distance.
The algorithm is simple and straightforward: if the measured distance is
greater than the predicted distance, the stability is negative, and if the
measured distance is less than the predicted distance, the stability is
positive. The stability value is normalized to be between -1 and 1, where -1
indicates the least stable (measured distance is much greater than predicted)
and 1 indicates the most stable (measured distance is much less than predicted).
}
\examples{
calcStability(predicted.dist = 0.3, measured.dist = 0.5)
calcStability(predicted.dist = 0.3, measured.dist = 0.1)

}
