% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamPred.R
\name{gamPred}
\alias{gamPred}
\title{calculation of stability using an generalized additive model.}
\usage{
gamPred(
  comtable,
  envmeta,
  comdist = NULL,
  sitenames = NULL,
  GAM.dist.method = "manhattan"
)
}
\arguments{
\item{comtable}{The community table}

\item{envmeta}{The environmental metadata table/matrix}

\item{comdist}{The community dissimilarity matrix (optional, default: NULL)}

\item{sitenames}{The names of the site (optional, default: NULL)}

\item{GAM.dist.method}{The method for calculating dist (default: "manhattan")}
}
\value{
a column vector of predicted stability values for each site
}
\description{
TODO: fill this description
This function ......
}
\examples{
varespec <- data(varespec)
varechem <- data(varechem)
example.stability_GAM <- gamPred(varespec, varechem)

}
