% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmPred.R
\name{gdmPred}
\alias{gdmPred}
\title{calculation of stability using a (nonlinear) generalized dissimilarity model.}
\usage{
gdmPred(comdist, envmeta, sitenames = NULL, X = NULL, Y = NULL)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envmeta}{The environmental metadata table/matrix}

\item{sitenames}{The names of the site (optional, default: NULL)}

\item{X}{The X coordinates of the sites (optional, default: NULL)}

\item{Y}{The Y coordinates of the sites (optional, default: NULL)}
}
\value{
a column vector of predicted stability values for each site
}
\description{
TODO: fill this description
This function ......
}
\examples{
varespec <- data(varespec)
varechem <- data(varechem)
example.comdist <- betaDiv(varespec)
example.stability_GDM <- gdmPred(example.comdist, varechem)

}
