% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPred.R
\name{glmPred}
\alias{glmPred}
\title{calculation of stability using a generalized linear model.}
\usage{
glmPred(comdist, envmeta, sitenames = NULL)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envmeta}{The environmental metadata table/matrix}

\item{sitenames}{The names of the site}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the dissimilarity matrix and the environmental
matrix as input, and calculate the stability of each site using a generalized
linear model (gLM), where the contributions are constrained as non-negative
\code{lower.limits=0} to ensure the explanability of each coefficient.
The stability is calculated by comparing the predicted distance (based on the
linear model) and the mean measured distance (based on vegdist function).
}
\examples{
library(vegan)
data(varespec)
data(varechem)
example.comdist <- vegdist(varespec)
example.stability_GLM <- glmPred(example.comdist, varechem)

}
