% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearPred.R
\name{linearPred}
\alias{linearPred}
\title{calculation of stability using linear prediction model.}
\usage{
linearPred(comdist, envdist, sitenames = NULL)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envdist}{The environmental dissimilarity matrix}

\item{sitenames}{The names of the site}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the diversity matrix and the environmental distance
matrix as input, and calculate the stability of each site using linear model.
The stability is calculated by comparing the predicted distance (based on the
linear model) and the mean measured distance (based on vegdist function).
}
\examples{
library(vegan)
library(stats)
data(varespec)
data(varechem)
example.comdist <- vegdist(varespec)
example.envdist <- dist(
    BBmisc::normalize(
        varechem,
        method = "range",
        margin = 2
    ),
    method = "euclidean"
)
example.stability_LM <- linearPred(example.comdist, example.envdist)

}
