% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlPred.R
\name{mlPred}
\alias{mlPred}
\title{calculation of stability using multiple linear regression model.}
\usage{
mlPred(comdist, envmeta, sitenames = NULL)
}
\arguments{
\item{comdist}{The community dissimilarity matrix}

\item{envmeta}{The environmental metadata table/matrix}

\item{sitenames}{The names of the site}
}
\value{
a column vector of predicted stability values for each site
}
\description{
This function will take the diversity matrix and the environmental distance
matrix as input, and calculate the stability of each site using multiple
linear model (ML). The stability is calculated by comparing the predicted
distance (based on the multiple linear model) and the mean measured distance
between the site and other sites (based on the difference of envmeta and
the corresponding comdist).
}
\examples{
data(varespec)
data(varechem)
example.comdist <- betaDiv(varespec)
example.stability_ML <- mlPred(example.comdist, varechem)

}
